requirements.txt
dj-database-url==2.3.0
Django==4.2.18
django-forms-bootstrap==3.1.0
django-netfields==1.3.2
django-pglocks==1.0.4
djangorestframework==3.15.2
djangorestframework-csv==3.0.2
gunicorn==23.0.0
ipaddress==1.0.23
netaddr==1.3.0
psycopg==3.2.4
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
whitenoise==6.8.2